Move section title to homepage module
